senoko-shell: Decrease heap size and priority level
The shell is a low-priority thread.  Lower it down so that it's above
only the main thread, idle thread, and watchdog thread.  This allows
the charger and gas gauge monitoring functions to take priority.

Additionally, decrease the stack size to 1024 bytes, since we didn't
need to take a full fifth of the RAM of the system for a simple shell.

Signed-off-by: Sean Cross <xobs@kosagi.com>