senoko: chconf: Set periodic ticks, add senoko_uptime
In order to be more precise, we now use periodic ticks rather than
freerunning mode.  This may not actually affect anything, and may
not be a good idea, but it's what we're running with now.

At 1000 Hz, the timer will tick and we'll increment the senoko_uptime
variable.  This gives us more bits to work with, and prevents the
uptime from overflowing every 65 seconds.

Signed-off-by: Sean Cross <xobs@kosagi.com>