chg: Add slop for capacity detection
We detect boards that need programming by looking at the "capacity" field.
Sometimes the mainboard will ask to view data in mWh instead of mAh, in
which case the numbers will be different.  Also, due to voltages and
rounding, the capacity can change.

Add a slop value so that we read a range of values.

Signed-off-by: Sean Cross <xobs@kosagi.com>