STM32: i2c_lld: Fix i2c slave restart behavior
When performing random reads, multiple start conditions are encountered
before a stop condition.  If this happens, txcount or rxcount will be
nonzero.

Ordinarily, this would cause the current buffer to get dropped.  Modify
the i2c LLD code to instead call the callback (if one exists) if we
get a STARTF condition when data has been received.

Signed-off-by: Sean Cross <xobs@kosagi.com>