chg: Remove FET manipulation code
Manipulating the FETs directly didn't seem like a good idea.  After reading
up on it, it looks like we have to manually enable ImpedenceTrack in order
to get the FETs to work properly.

Now that the gas gauge is calibrated and we're running ImpedenceTrack, it
should be able to control the FETs on its own.

Signed-off-by: Sean Cross <xobs@kosagi.com>