chg: Run IT command when we ought to be able to charge
When both the CHG and DSG FETs are off and the system is discharging,
this is generally a sign that we ought to be able to actually charge.
The solution here is to run the IT command.

It is unclear what this actually does, but it seems to kick it into a
mode where it actually works.

This is particularly useful when doing a cold-start.

Signed-off-by: Sean Cross <xobs@kosagi.com>