senoko-i2c: Add a thread to unstick I2C
I2C will sometimes get in a state where SDA is forced low by the STM32,
even when it hasn't been talked to.  No interrupts are generated in such
a situation, so there's nothing to listen for.

Instead, we start a thread that polls every 20 ms.  If it sees the bus
being busy more than four times in a row, it will reset the I2C bus and
hopefully unstick it.

Signed-off-by: Sean Cross <xobs@kosagi.com>