power: force board on or off during a reboot
Sometimes, Senoko will reboot and get the wrong power state, and it
will fail to set the power state due to being in a "cooldown" mode.

Since the pin defaults to On, this can sometimes result in a Senoko that
thinks it's powered off, while it's actually powered on.

Forcing the power_off or power_on state can help us sync reality with
what Senoko believes is the case.

Signed-off-by: Sean Cross <xobs@kosagi.com>